requirements.txt
# Core Django
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.10
redis==5.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
django-environ==0.11.2

# Async Task Queue
celery==5.3.1

# AI & Audio Processing
openai==1.3.0
anthropic==0.7.0
requests==2.31.0

# API Documentation
drf-spectacular==0.26.4

# Development Tools
django-extensions==3.2.3
ipython==8.15.0

# Production Server
gunicorn==21.2.0
whitenoise==6.5.0

# Testing
pytest==7.4.2
pytest-django==4.5.2
pytest-asyncio==0.21.1
coverage==7.3.0
factory-boy==3.3.0

# Monitoring & Logging
django-debug-toolbar==4.2.0
sentry-sdk==1.32.0

# Utilities
Pillow==10.0.1
python-dateutil==2.8.2
